// src/lib/drive/api/types.ts
export interface FileListItem {
  category?: string
  content_hash?: string
  file_extension?: string
  created_at: string
  drive_id: string
  file_id: string
  mime_type?: string
  name: string
  parent_file_id: string
  punish_flag?: number
  size?: number
  starred: boolean
  sync_device_flag: boolean
  sync_flag: boolean
  sync_meta: string
  type: 'file' | 'folder'
  updated_at: string
  url: string
  user_meta?: string
  user_tags?: Record<string, string>
}
// ...
export interface FileGetPathItem {
  trashed: boolean
  drive_id: string
  file_id: string
  created_at: string
  domain_id: string
  encrypt_mode: string
  hidden: boolean
  name: string
  parent_file_id: string
  starred: boolean
  status: string
  type: 'file' | 'folder'
  updated_at: string
  sync_flag: boolean
}